Adding "locale" to readme
I tried to ctrl+F the readme for "locale" to find that term and thought that maybe this library didn't support switching locales. It was only after reading #80 that I realized `domain` was the parameter to set. I thought explicitly adding "locale" to the docs might save someone time in the future.